move prefetching storage in separate module
